+++ NEWS.md
@@ -13,6 +13,8 @@ Changes:
 
 5. All `DatabaseConnector` unique top-level functions (e.g. `querySql()` and `getTableNames()`) now work on all `DBI` connections, not just `DatabaseConnector` connections. This is for improved separation of code, and should make it easier to add and remove new database platforms in the future.
 
+6. On connecting to DuckDB checks if the ICU extension is installed and if not, attempts to install it. The ICU extension is needed for some date and time functions such as `CURRENT_DATE`.
+
 
 Bugfixes: